Move types to sidebar
